+++ classes/output/edit_renderer.php
@@ -215,7 +215,7 @@ class edit_renderer extends \plugin_renderer_base {
      * @param \stdClass $offlinequiz
      * @param \stdClass $cm
      * @param array $templatecontext
-     * @return string
+     * @return array
      */
     public function offlinequiz_group_selector_values($offlinequiz, $cm, $templatecontext) {
 
@@ -243,7 +243,7 @@ class edit_renderer extends \plugin_renderer_base {
      *
      * @param structure $structure the offlinequiz structure.
      * @param array $templatecontext
-     * @return string HTML to output.
+     * @return array HTML to output.
      */
     public function offlinequiz_state_warnings_values(structure $structure, $templatecontext) {
 
@@ -268,7 +268,7 @@ class edit_renderer extends \plugin_renderer_base {
      *
      * @param structure $structure the offlinequiz structure.
      * @param array $templatecontext
-     * @return string HTML to output.
+     * @return array HTML to output.
      */
     public function offlinequiz_information_values(structure $structure, $templatecontext) {
 
@@ -285,7 +285,7 @@ class edit_renderer extends \plugin_renderer_base {
      * @param \stdClass $offlinequiz the offlinequiz settings from the database.
      * @param \moodle_url $pageurl the canonical URL of this page.
      * @param array $templatecontext
-     * @return string HTML to output.
+     * @return array HTML to output.
      */
     public function maximum_grade_input_values($offlinequiz, \moodle_url $pageurl, $templatecontext) {
         $templatecontext['maxgradeinputsize'] = $offlinequiz->decimalpoints + 2;
@@ -299,7 +299,7 @@ class edit_renderer extends \plugin_renderer_base {
      * @param \moodle_url $pageurl the canonical URL of this page.
      * @param \stdClass $offlinequiz
      * @param array $templatecontext
-     * @return string HTML to output.
+     * @return array HTML to output.
      */
     protected function repaginate_button_values(structure $structure, \moodle_url $pageurl, $offlinequiz, $templatecontext) {
 
@@ -426,7 +426,7 @@ class edit_renderer extends \plugin_renderer_base {
      *
      * @param \stdClass $offlinequiz the offlinequiz settings from the database.
      * @param array $templatecontext
-     * @return string HTML to output.
+     * @return array HTML to output.
      */
     public function total_marks($offlinequiz, $templatecontext) {
         $templatecontext['sumgrades'] = offlinequiz_format_grade($offlinequiz, $offlinequiz->sumgrades);
@@ -783,6 +783,8 @@ class edit_renderer extends \plugin_renderer_base {
                                 $this->random_question($structure, $question, $pageurl) :
                                 $this->question_name($structure, $question, $pageurl)
                             ),
+            'bankname' => $this->get_bank_name($structure, $question),
+            'bankurl' => $this->get_bank_url($structure, $question),
             'questionpreviewicon' => $this->question_preview_icon($structure->get_offlinequiz(), $question),
             'questionremoveicon' => ($structure->can_be_edited() ? $this->question_remove_icon($question, $pageurl) : ''),
             'questionmarkicon' => $this->marked_out_of_field($structure->get_offlinequiz(), $question),
@@ -829,6 +831,8 @@ class edit_renderer extends \plugin_renderer_base {
                                 $this->random_question($structure, $question, $pageurl) :
                                 $this->question_name($structure, $question, $pageurl)
                             ),
+            'bankname' => $this->get_bank_name($structure, $question),
+            'bankurl' => $this->get_bank_url($structure, $question),
             'questionpreviewicon' => $this->question_preview_icon($structure->get_offlinequiz(), $question),
             'questionremoveicon' => ($structure->can_be_edited() ? $this->question_remove_icon($question, $pageurl) : ''),
             'questionmarkicon' => $this->marked_out_of_field($structure->get_offlinequiz(), $question),
@@ -848,6 +852,32 @@ class edit_renderer extends \plugin_renderer_base {
         $output .= html_writer::end_tag('div');
 
         return $output;
+    }
+    /**
+     * get the url of the question bank
+     * @param mixed $structure
+     * @param mixed $question
+     * @return string
+     */
+    private function get_bank_url($structure, $question) {
+        $bankurl = new \moodle_url('/question/edit.php',
+                [
+                    'cmid' => $structure->get_source_bank_cminfo($question->contextid)->id,
+                    'cat' => "{$question->categoryid},{$question->contextid}",
+                ]
+            );
+        return $bankurl->out(false);
+    }
+    /**
+     * get the name of the question bank
+     * @param mixed $structure
+     * @param mixed $question
+     */
+    private function get_bank_name($structure, $question) {
+        if ($structure->get_context()->id != $question->contextid) {
+            return $question->qcname;
+        }
+        return '';
     }
 
     /**