+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 -----------------
 * [Feature] 5.1 compatible version
 * [Feature] added a setting to set a default max mark
+* [Feature] Display question
 * [Bug] Changed that partial results also show the name and user profile picture
 * [Bug] Support underline formatting of tinymce